document-extractor-api/requirements.txt
# Core Flask dependencies
flask[async]==3.1.*
flask-cors==5.0.*

# AI Integration
openai==1.55.3
pydantic==2.9.*

# Data Processing (CSV/Excel)
pandas==2.2.*
openpyxl==3.1.*

# File & Environment Handling
python-multipart
python-dotenv
pillow

# Development Tools
ruff
black
pytest

# Optional: For better async support
asyncio-throttle==1.0.*